Adjust the PomodoroApp component
Change the alert from within the componentDidMount method to a function call that triggers a sound.

Create a method called playAlarmSound to be called when the pomodoro timer runs out.

Change the texts of the project's button elements, the first button will have the text Stop, the second button will have the text Start and the third button will have the text Reset.

Add the audio tag with the alarm sound inside the component's render method.